tpm2: Add a TPMv2 MMIO TIS driver
Add a TPMv2 TIS MMIO compatible driver.
This useful for a couple of reasons.  First of all we can support a TPMv2
on devices that have this hardware (e.g the newly added SynQuacer box).
We can also use the driver in our QEMU setups and test the EFI TCG2
protocol, which cannot be tested with our sandbox.

Ideally we should create an abstraction for all TIS compatible TPMs we
support. Let's plug in the mmio driver first.

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>