hack: Enable RPMB emulation in optee supplicant
StMM functionality storing variables on an RPMB is fragile, mostly because
it depends on 3 different projects (U-Boot, OP-TEE and EDK2).  In order to
provide a U-Boot we can use for our OP-TEE CI, hack the mmc layer and
provide RPMB emulation similar to the linux supplicant.
Note that proper RPMB hardware support is broken with this patch applied.

A defconfig (qemu_tfa_mm_defconfig) is provided to make the build process
easier.

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>